:art: Minor cleanup
drf-yasg performs pretty much a module-level import of the schema info
object as part of the generate-swagger command. This, in turn, causes
the notifs documentation to be generated BEFORE the urlconf is loaded,
which imports the viewsets and registers the resources and actions.

Importing the viewsets right after django starts up (and thus before
management commands run) ensures the viewset metaclass ran, and the
resources and actions are registered.