:sparkles: Added filter option maximaleVertrouwelijkheidaanduiding
Vertrouwelijkheidaanduiding is ordered from more to less public. Using
the maximaleVertrouwelijkheidaanduiding parameter will return zaken
that are more public or equally public to the specified VA.

This removes the needs for client to perform client-side filtering of
results, ensures that only data is exchanged that clearance is given
for, and improves performance considerably because there is no
pagination to go through (sequentially) on the client side.