:bug: Fix failing tests
drf-nested-viewsets is injecting URL kwargs into the request.data, which
breaks the explicit validation error when the zaak is attempted to be
modified.

In normal circumstances, that would be fine, but we can't introduce
API changes in this version. Perhaps in the future the zaak attribute
will be dropped from the body since it's implied from the URL anyway,
which would make it read-only.